Adjust function max_intersections code structure, optimize max_intersection performance, and add test cases
