+++ extension/core_functions/aggregate/holistic/CMakeLists.txt
@@ -6,7 +6,8 @@ add_library_unity(
   reservoir_quantile.cpp
   mad.cpp
   approximate_quantile.cpp
-  mode.cpp)
+  mode.cpp
+  max_intersections.cpp)
 set(CORE_FUNCTION_FILES
     ${CORE_FUNCTION_FILES} $<TARGET_OBJECTS:duckdb_core_functions_holistic>
     PARENT_SCOPE)
