+++ extension/core_functions/include/core_functions/aggregate/distributive_functions.hpp
@@ -299,14 +299,4 @@ struct SumNoOverflowFun {
 	static AggregateFunctionSet GetFunctions();
 };
 
-struct MaxIntersectionsFun {
-	static constexpr const char *Name = "max_intersections";
-	static constexpr const char *Parameters = "left,right";
-	static constexpr const char *Description = "Returns the maximum number of overlapping intervals from pairs of (left, right) values.";
-	static constexpr const char *Example = "max_intersections(left, right)";
-	static constexpr const char *Categories = "";
-
-	static AggregateFunction GetFunction();
-};
-
 } // namespace duckdb
